requirements.txt
ffmpeg-python==0.2.0
future==1.0.0
typing_extensions==4.15.0
yt-dlp==2026.2.21
